palomar: more cleanup
